+++ src/plugins/gmedecoder/GmeIndexerSource.cpp
@@ -150,9 +150,9 @@ void GmeIndexerSource::UpdateMetadata(
 {
     /* only need to do this check once, and it's relatively expensive because
     it requires a db read. cache we've already done it. */
-    int modifiedTime = fs::getLastModifiedTime(fn);
+    int64_t modifiedTime = fs::getLastModifiedTime(fn);
     const std::string firstExternalId = indexer::createExternalId(EXTERNAL_ID_PREFIX, fn, 0);
-    int modifiedDbTime = indexer->GetLastModifiedTime(this, firstExternalId.c_str());
+    int64_t modifiedDbTime = indexer->GetLastModifiedTime(this, firstExternalId.c_str());
     if (modifiedDbTime < 0 || modifiedTime != modifiedDbTime) {
         fn = fs::canonicalizePath(fn);
